requirements.txt
python-docx>=0.8.11
gradio>=4.0.0
ollama>=0.1.0
